+++ test/AutoMapper.Contrib.Autofac.DependencyInjection.TestInfrastructure/Resolver/CustomerTypeConverter.cs
@@ -3,6 +3,7 @@ using AutoMapper.Contrib.Autofac.DependencyInjection.TestInfrastructure.Entities
 
 namespace AutoMapper.Contrib.Autofac.DependencyInjection.TestInfrastructure.Resolver
 {
+    // ReSharper disable once ClassNeverInstantiated.Global
     public class CustomerTypeConverter : ITypeConverter<CustomerDto, Customer>
     {
         public Customer Convert(CustomerDto source, Customer destination, ResolutionContext context)