+++ test/AutoMapper.Contrib.Autofac.DependencyInjection.Tests/ContainerBuilderExtensionTests.cs
@@ -22,7 +22,12 @@ namespace AutoMapper.Contrib.Autofac.DependencyInjection.Tests
             var mapper = container.Resolve<IMapper>();
 
             var customerDto = mapper.Map<CustomerDto>(customer);
+            var reverseMappedCustomer = mapper.Map<Customer>(customerDto);
 
+            Assert.Equal(customer.Id, reverseMappedCustomer.Id);
+            Assert.Equal(customer.FirstName, reverseMappedCustomer.FirstName);
+            Assert.Equal(customer.Name, reverseMappedCustomer.Name);
+            
             Assert.Equal(customer.Id, customerDto.Id);
             Assert.Equal(customer.FirstName, customerDto.FirstName);
             Assert.Equal(customer.Name, customerDto.Name);
@@ -38,6 +43,7 @@ namespace AutoMapper.Contrib.Autofac.DependencyInjection.Tests
 
             Assert.True(container.IsRegistered<IEnumerable<Profile>>());
             Assert.True(container.IsRegistered<MapperConfiguration>());
+            Assert.True(container.IsRegistered<IConfigurationProvider>());
             Assert.True(container.IsRegistered<IMapper>());
             Assert.True(container.IsRegistered<IValueResolver<Customer, CustomerDto, string>>());
             Assert.True(container.IsRegistered<ITypeConverter<CustomerDto, Customer>>());